Remove the concept of environment from the sys package as well as sys.Init checks for the env and migration_path.
The "env" var was only used for setting a default value for the gin debug mode and the ErrorIfProd func.  We already support the http_debug env var and the ErrorIfProd func was only used for migrations (which have been removed) and preventing seeders from running in production (in some apps).

Application logic should not changed based on the environment it is running in.  Environment behavior should be changed through configuration, not code.  For this reason, having the environment built into framework code seemed unnecessary.